requirements.txt
matplotlib~=3.7.1
numpy==1.24.2
matplotlib==3.7.1
numpy-quaternion==2022.4.3